Create README.md
Tixon is an experimental digital currency that enables instant payments to anyone, anywhere in the world. Tixon uses peer-to-peer technology to operate with no central authority: managing transactions and issuing money are carried out collectively by the network. Tixon Core is the name of open source software which enables the use of this currency.

For more information read the original Bitcoin whitepaper.